+++ pikoci/mysql/migrate/migrations/migrations.go
@@ -16,7 +16,7 @@ type Migration struct {
 // Migrations is the ordered list of all schema migrations. It is defined as
 // a fixed-size array so that the compiler catches ordering conflicts when
 // multiple developers add migrations concurrently.
-var Migrations = [51]Migration{
+var Migrations = [52]Migration{
 	V0Initial,
 	V1ResourceCheckInterval,
 	V2JobsAndBuilds,
@@ -68,4 +68,5 @@ var Migrations = [51]Migration{
 	V48DisableRetry,
 	V49AllowFailure,
 	V50Interruptible,
+	V51JobInputs,
 }